+++ scripts/hotWalletCli/logging.ts
